requirements.txt
aws-cdk.core==1.128.0
aws-cdk.aws_events==1.128.0
aws-cdk.aws_events_targets==1.128.0
aws-cdk.aws_lambda==1.128.0
aws-cdk.aws_stepfunctions==1.128.0
aws-cdk.aws_stepfunctions_tasks==1.128.0
pyyaml<5.5, >=3.10